Start rework of motion ID system to 'timestamp offset' system
Structure refactor and naming changes, small prep work.

+++ firmware/src/app_state_machines/app_events.h
@@ -95,7 +95,7 @@ typedef struct ExpansionServoRequestEvent__
 typedef struct BarrierSyncEvent__
 {
     StateEvent super;    // Encapsulated event reference
-    uint16_t   id;       // UUID of the movement/lighting that we want to synchronise
+    uint32_t   id;       // UUID of the movement/lighting that we want to synchronise
 } BarrierSyncEvent;
 
 /* -------------------------------------------------------------------------- */
